requirements.txt
yfinance==1.2.0
polygon-api-client==1.16.3
requests==2.32.5
python-dotenv==1.2.1
streamlit==1.54.0
plotly==6.5.2
pandas==2.3.3
numpy==2.2.6
vaderSentiment==3.3.2
pytrends==4.9.2
beautifulsoup4==4.14.3
lxml==6.0.2
